requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=py37h06a4308_0
aiohttp=3.8.3=py37h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
alembic=1.8.1=py37h06a4308_0
apex=0.1=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
asn1crypto=1.5.1=py37h06a4308_0
async-timeout=4.0.2=py37h06a4308_0
asynctest=0.13.0=py_0
attrs=22.1.0=py37h06a4308_0
blas=1.0=mkl
blessed=1.19.1=pypi_0
blinker=1.4=py37h06a4308_0
boto3=1.26.27=pyhd8ed1ab_0
botocore=1.29.27=pyhd8ed1ab_0
bottleneck=1.3.5=py37h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.12.7=ha878542_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
configparser=5.0.2=pyhd3eb1b0_0
cryptography=38.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.32=pypi_0
databricks-cli=0.17.3=pyhd8ed1ab_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
docker-py=4.4.1=py37h06a4308_5
docker-pycreds=0.4.0=pyhd3eb1b0_0
entrypoints=0.4=py37h06a4308_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.8.0=pypi_0
flask=2.1.3=py37h06a4308_0
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py37h5eee18b_0
gensim=3.8.3=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
gpustat=1.0.0=pypi_0
greenlet=2.0.1=py37h6a678d5_0
grpcio=1.42.0=py37hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gunicorn=20.1.0=py37h06a4308_0
huggingface_hub=0.11.1=py_0
icu=58.2=he6710b0_3
idna=3.4=py37h06a4308_0
importlib-metadata=5.1.0=pypi_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2022.1.0=h9e868ea_3769
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py37h06a4308_0
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
mako=1.2.3=py37h06a4308_0
markdown=3.3.4=py37h06a4308_0
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.5.2=py37h06a4308_0
matplotlib-base=3.5.2=py37hf590b9c_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
mlflow=1.30.0=py37h02d9ccd_0
multidict=6.0.2=py37h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
nlg-eval=2.3=pypi_0
nltk=3.7=pyhd3eb1b0_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.7.3=py37hb2eb853_0
numpy=1.21.6=pypi_0
numpy-base=1.19.2=py37hfa32c7d_0
nvidia-ml-py=11.495.46=pypi_0
oauthlib=3.2.1=py37h06a4308_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h8c16a72_0
pcre=8.45=h295c915_0
pillow=9.2.0=py37hace64e9_1
pip=22.2.2=py37h06a4308_0
ply=3.11=py37_0
prometheus_client=0.14.1=py37h06a4308_0
prometheus_flask_exporter=0.21.0=pyhd8ed1ab_0
protobuf=3.20.1=pypi_0
psutil=5.9.4=pypi_0
py-rouge=1.1=pyhd8ed1ab_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.4.0=py37h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pysocks=1.7.1=py37_1
python=3.7.15=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py37h06a4308_0
pyyaml=6.0=py37h5eee18b_1
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
querystring_parser=1.2.4=py_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.6.0=py37h06a4308_0
sacremoses=0.0.53=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.4.1=py37h0b6359f_0
sentencepiece=0.1.96=py37h7cecad7_1
setuptools=59.5.0=pypi_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.2.0=pypi_0
smmap=4.0.0=pyhd3eb1b0_0
sqlalchemy=1.4.39=py37h5eee18b_0
sqlite=3.40.0=h5082296_0
sqlparse=0.4.3=py37h06a4308_0
tabulate=0.8.10=py37h06a4308_0
tensorboard=2.10.0=py37h06a4308_0
tensorboard-data-server=0.6.0=py37hca6d32c_0
tensorboard-plugin-wit=1.8.1=py37h06a4308_0
tensorboardx=2.5.1=pypi_0
theano=1.0.5=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torch=1.10.1+cu111=pypi_0
torchaudio=0.11.0=py37_cu113
torchvision=0.12.0=py37_cu113
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=py37h06a4308_0
transformers=4.2.2=pypi_0
typing-extensions=4.4.0=py37h06a4308_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.13=py37h06a4308_0
wcwidth=0.2.5=pypi_0
websocket-client=0.58.0=py37h06a4308_4
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xdg=5.1.1=pypi_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py37h5eee18b_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
